+++ python/ray/rllib/optimizers/__init__.py
@@ -2,8 +2,9 @@ from ray.rllib.optimizers.async import AsyncOptimizer
 from ray.rllib.optimizers.local_sync import LocalSyncOptimizer
 from ray.rllib.optimizers.multi_gpu import LocalMultiGPUOptimizer
 from ray.rllib.optimizers.sample_batch import SampleBatch
+from ray.rllib.optimizers.evaluator import Evaluator, TFMultiGPUSupport
 
 
 __all__ = [
     "AsyncOptimizer", "LocalSyncOptimizer", "LocalMultiGPUOptimizer",
-    "SampleBatch"]
+    "SampleBatch", "Evaluator", "TFMultiGPUSupport"]
